6-size.c

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ int main(void)
 	float f;
 	long long int lli;
 
-	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(c));
-	printf("Size of a int: %lu byte(s)\n", (unsigned long)sizeof(i));
-	printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(li));
-	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
-	printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(lli));
+	printf("Size of a char: %zu byte(s)\n", sizeof(c));
+	printf("Size of a int: %zu byte(s)\n", sizeof(i));
+	printf("Size of a long int: %zu byte(s)\n", sizeof(li));
+	printf("Size of a float: %zu byte(s)\n", sizeof(f));
+	printf("Size of a long long int: %zu byte(s)\n", sizeof(lli));
 	return (0);
 }
